+++ src/core/App.js
@@ -4,9 +4,9 @@ import {
   Route,
   Switch,
 } from "react-router-dom/cjs/react-router-dom";
-import NavigationBar from "./common/NavigationBar/index";
-import MainPage from "./features/MainPage";
-import PersonList from "./features/PersonList";
+import NavigationBar from "../common/NavigationBar";
+import MainPage from "../features/MainPage";
+import PersonList from "../features/PersonList";
 import { toMainPage, toPersonList } from "./routes";
 
 function App() {
