+++ src/index.js
@@ -2,10 +2,10 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import { ThemeProvider } from "styled-components";
-import { theme } from "./theme";
-import App from "./App";
-import { GlobalStyle } from "./GlobalStyle";
-import store from "./store";
+import { theme } from "./core/theme";
+import App from "./core/App";
+import { GlobalStyle } from "./core/GlobalStyle";
+import store from "./core/store";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
