--- src/rootSaga.js
@@ -1,6 +1,0 @@
-import { all } from "redux-saga/effects";
-import { watchFetchPopularMovies } from "./features/MainPage/popularMoviesSaga";
-
-export default function* rootSaga() {
-  yield all([watchFetchPopularMovies()]);
-}
