--- src/store.js
@@ -1,20 +1,0 @@
-import { configureStore } from "@reduxjs/toolkit";
-import createSagaMiddleware from "@redux-saga/core";
-import popularMoviesReducer from "./features/MainPage/popularMoviesSlice";
-import paginationReducer from "./common/Pagination/paginationSlice";
-import rootSaga from "./rootSaga";
-
-const sagaMiddleware = createSagaMiddleware();
-
-const store = configureStore({
-  reducer: {
-    popularMovies: popularMoviesReducer,
-    pagination: paginationReducer,
-  },
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware().concat(sagaMiddleware),
-});
-
-sagaMiddleware.run(rootSaga);
-
-export default store;
